update and add new exception

+++ src/main/java/hexlet/code/demo/service/TaskStatusService.java
@@ -4,15 +4,16 @@ import hexlet.code.demo.dto.TaskStatusCreateDTO;
 import hexlet.code.demo.dto.TaskStatusDTO;
 
 import hexlet.code.demo.dto.TaskStatusUpdateDTO;
+import hexlet.code.demo.exception.ResourceAssociationException;
+import hexlet.code.demo.exception.ResourceNotFoundException;
 import hexlet.code.demo.mapper.TaskStatusMapper;
 
 import hexlet.code.demo.repository.TaskStatusRepository;
 
-import jakarta.persistence.EntityNotFoundException;
-
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -24,19 +25,22 @@ public class TaskStatusService {
 
     private final TaskStatusMapper taskStatusMapper;
 
+    @Transactional(readOnly = true)
     public TaskStatusDTO getTaskStatusById(Long id) {
         var taskStatus = taskStatusRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Task status with id " + id + " not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Task status with id " + id + " not found"));
         return taskStatusMapper.toDTO(taskStatus);
     }
 
-    public List<TaskStatusDTO> getTaskStatusAll() {
+    @Transactional(readOnly = true)
+    public List<TaskStatusDTO> getAllTaskStatuses() {
         var taskStatuses = taskStatusRepository.findAll();
         return taskStatuses.stream()
                 .map(taskStatusMapper::toDTO)
                 .toList();
     }
 
+    @Transactional
     public TaskStatusDTO createTaskStatus(TaskStatusCreateDTO dto) {
         var taskStatus = taskStatusMapper.toEntity(dto);
 
@@ -45,9 +49,10 @@ public class TaskStatusService {
         return taskStatusMapper.toDTO(taskStatus);
     }
 
+    @Transactional
     public TaskStatusDTO updateTaskStatus(Long id, TaskStatusUpdateDTO dto) {
         var taskStatus = taskStatusRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Task status with id " + id + " not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Task status with id " + id + " not found"));
 
         taskStatusMapper.updateEntityFromDTO(dto, taskStatus);
 
@@ -56,9 +61,14 @@ public class TaskStatusService {
         return taskStatusMapper.toDTO(taskStatus);
     }
 
+    @Transactional
     public void deleteTaskStatus(Long id) {
         var taskStatus = taskStatusRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Task status with id " + id + " not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Task status with id " + id + " not found"));
+
+        if (!taskStatus.getTasksList().isEmpty()) {
+            throw new ResourceAssociationException("Task status with id " + id + " is associated with tasks");
+        }
 
         taskStatusRepository.delete(taskStatus);
     }
